+++ .env.development
@@ -8,3 +8,6 @@ NEXT_PUBLIC_APP_URL=http://localhost:3000
 # you can add it to .env.local but it will then disable the loading in the local
 # preview as well.
 NEXT_PUBLIC_DISABLE_LOADING=false
+
+# UI Lab URL
+NEXT_PUBLIC_UI_LAB_URL=http://localhost:3001
