+++ features/UiLab/index.ts
@@ -1,1 +1,2 @@
 export * from './Page';
+export * from './Box';
