+++ lib/config.ts
@@ -156,6 +156,7 @@ export const PRODUCTION_LIVE =
 export const PREVIEW = process.env.NEXT_PUBLIC_VERCEL_ENV === 'preview';
 export const BUILD_DATE = process.env.NEXT_PUBLIC_BUILD_DATE as string;
 export const GIT_COMMIT_SHA = process.env.NEXT_PUBLIC_VERCEL_GIT_COMMIT_SHA;
+export const UI_LAB_URL = process.env.NEXT_PUBLIC_UI_LAB_URL as string;
 export const APP_URL =
   process.env.NEXT_PUBLIC_APP_URL ??
   (PREVIEW
