+++ package.json
@@ -7,7 +7,6 @@
   "scripts": {
     "analyze": "ANALYZE=true bun run build",
     "build": "next build",
-    "build:update": "bun update @joonassandell/ui-lab && bun run build",
     "build:ignore": "IGNORE_ERRORS=true bun run build",
     "check": "bun run lint && bun run typecheck",
     "create:component": "bun --bun ./scripts/create-component.cjs",
@@ -20,7 +19,6 @@
     "typecheck": "tsc --pretty --noEmit"
   },
   "dependencies": {
-    "@joonassandell/ui-lab": "^0.1.7",
     "@next/third-parties": "^14.2.4",
     "@nivo/line": "^0.86.0",
     "clsx": "^2.1.1",
