+++ stylesheets/generic/base.scss
@@ -70,7 +70,7 @@ img::-moz-selection {
   margin: 0;
 }
 
-:where(img, video) {
+:where(img, video, iframe) {
   max-width: 100%;
 }
 
@@ -88,4 +88,9 @@ img::-moz-selection {
   border-collapse: collapse;
   border-color: inherit;
   text-indent: 0;
+}
+
+:where(iframe) {
+  background: transparent;
+  border: 0;
 }
